Module confertus::dynamic_vector

Actual implementation of dynamic bit vector with AVL Tree

Structs

DynamicBitVector

Implementation of Dynamic Bit Vector using AVL/RB tree.

Leaf

Leaf element of crate::DynamicBitVector. Next to its value (LeafValue) and bits used inside (nums), it contains a reference to its parent crate::Node.

Node

Node element of super::DynamicBitVector. Contains references (indices) to parent Node, left and right subtrees, as well as nums, the number of used bits in the left subtree, ones the number of ones in the left subtree, and size, the total capacity of the current subtree.

Type Definitions

LeafValue

Primitive type used as bit container in Leaf. Probably u64 or u128.