Module confertus::traits

Trait definitions, particularly for StaticBitVec and DynBitVec.

Traits

BitSize

Trait to get instance bit size for different structs.

Dot

Visualize Tree-based structures with graphviz using the .dot format.

DynBitTree

DynBitVec

Functions associated with dynamic bit vectors.

StaticBitVec

Functions associated with static bit vectors. Not to be confused with specific containers such as u64, u128 or particulary Leaf, which additionally tracks the number of used bits, nums, and a parent Node.