Trait confertus::traits::DynBitVec

pub trait DynBitVec<T: StaticBitVec> {
    fn insert(self, i: usize, bit: bool);
    fn delete(self, i: usize);
    fn flip(self, i: usize);
    fn bitset(self, i: usize);
    fn bitclear(self, i: usize);
}

Required Methods

fn insert(self, i: usize, bit: bool)

insert i [0|1] insert a 0 or 1 at the i-th position of the bit vector concurrently updates all relevant ones and num values when traversing to location i, rebalance if necessary

fn delete(self, i: usize)

delete i delete the i-th bit concurrently updates all relevant ones and num values when traversing to location i, rebalancing if necessary

fn flip(self, i: usize)

flip i flip the i-th bit updates ones and num accordingly

fn bitset(self, i: usize)

bitset i sets i-th bit to 1 updates ones and num accordingly

fn bitclear(self, i: usize)

bitclear i sets i-th bit to 0 updates ones and num accordingly