Trait confertus::traits::StaticBitVec

pub trait StaticBitVec {
    type Intern;

    fn ones(&self) -> usize;
    fn access(&self, index: usize) -> bool;
    fn rank(&self, bit: bool, index: usize) -> usize;
    fn select(&self, bit: bool, n: usize) -> usize;
    fn values(&self) -> Self::Intern;
}

Functions associated with static bit vectors. Not to be confused with specific containers such as u64, u128 or particulary Leaf, which additionally tracks the number of used bits, nums, and a parent Node.

Required Associated Types

type Intern

Required Methods

fn ones(&self) -> usize

Return number of on-bits in container

fn access(&self, index: usize) -> bool

Access bit value at position index

fn rank(&self, bit: bool, index: usize) -> usize

Returns number of bit-values up to index in container

runtime complexity: O(1) to O(w)

fn select(&self, bit: bool, n: usize) -> usize

Return index of n-th bit-value in container

runtime complexity: O(1) to O(w)

fn values(&self) -> Self::Intern

Return full internal container